+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All changes to the MarkLogic (backend) portion of LUX capable of impacting the r
 
 ### Added
 
+- Added Part Of Work and Contains Work search terms to Works ([#234](https://github.com/project-lux/lux-marklogic/issues/234)).
+
 ### Changed
 
 ### Removed